+++ app/Data/VehicleDto.php
@@ -5,8 +5,8 @@ namespace App\Data;
 class VehicleDto
 {
     public function __construct(
-        public int $amount,
-        public int $balance,
+        public float $amount,
+        public float $balance,
         public int $due_date,
         public ?int $mileage = null,
     )
